get positions and properly allocate eigen memory in autodiff stack for vari_value<Eigen> and Eigen<double> operations
